Add files via upload
On MainActivity, enter the separator character and then click on Select .CSV file to open explorer and select your file. On successful operation on the file, click the button Preview to see the items on a ListView on the launched activity. On the bottom, click on the Write .CSV file to write and gives you at Toast message to the file location of the created file to where you can navigate.